We are now up to version 0.2.0 < build 14> release 14 or so...
This is quite a few updates in one... To sum it up I modified the autoloader a bit so the run commands are put into its own file, and I added a new feature to get by the blocking calls of include or return or whatever in some of the files... Basically an OnInit_<filename_excluding_.ahk>( ) function is called, if it exists, directly after the run commands.. This is so you can set up your default configuration settings per hotkey or whatever file because you wouldn't be able to otherwise without separating everything into separate files which wouldn't be needed or putting it into a folder which is included by the base-framework ( such as definitions ) but those are meant for other things... Definitions/ is for globals / ENUMeration / CONSTants, functions/ are for stand-alone functions only ( such as RunFuncIfExists which needed to be a function because if I injected if IsFunc( "name" ) { name( ) } it would see name( ) and claim it doesn't exist despite being inside if IsFunc which makes no sense - the language parser should be more intelligent but it isn't... ), classes/ are for libraries and classes / objects ( such as the math library which adds a lot of math functions and moves some from the global table, while leaving them there, into math.XXX to make it look cleaner -- and configuration -- etc.. I may go ahead and add a string library for quick / easy string functions such as string.safe and string.repeat, etc.. )....

Hotkeys added:
------------------------------------------------------------------------------------------------
Added math chars All using ALT + [ d adds shift option for delta case and some keys may require you to use shift to access depending on keyboard layout ] ( 0, d, +, -, <, >, =, *, / ) == ( ø / Δδ / ± / ± / ≤ / ≥ / ≡ / √ / ÷ )

Added German Umlauts  ALT + [ SHIFT for uppercase ] ( a / o / u / s ) == ( äÄ / öÖ / üÜ / ß )

Added currency chars and trademark / copyright symbols ALT + [ CTRL for c / l / e for other option or to avoid collisions ] ( c / l / t / e / r  ) == ( ©¢ / £ / € / ® / ™)

Some other notes:
------------------------------------------------------------------------------------------------
Removed the full path from hotkey_screencapture_current_window__alt_f12.ahk -- 1 more down and that may be all of them aside from the process but most of those I could replace using the config system...

Added a few libraries / classes, functions, placed all of the audible alerts into 1 file for Num/Scroll/CapsLock and it now uses configuration file and class for the sound - if unset it will set it.. you can alter the file in the config.ini file - more config options will appear soon.. all of them will add a default value and then you can alter it and reload the script to have the changes occur...

I will add a config option for the addons_exe folder so if you whitelist a file, it will not ask you again ( unless the file changed - I will use the hash of the file for security purposes ) for that file.

Note: When you first launch the framework, if you have any files inside the addons_exe folder, you'll be asked whether or not to include them.. Yes to include, no to skip it. For second load ( and any load where the load-order files exist ) you will be asked if you wish to load your current load-order or rebuild them.. Yes will launch as is ( default option ), no will rebuild the files which you'll need to do if new files are added to other folders.... I will eventually add in an auto-regenerate when new files are added, and I may make the addons_exe files that are ran a separately generated file so you won't be asked about those files when an auto-regeneration occurs...

Added German and other language Umlaut keys... so alt + a / o / u == äöü    AND alt + shift + a / o / u == Ä / Ö / Ü    for those of us using a keyboard layout that doesn't contain these keys, so we have easy access to upper and lowercase versions of them... I will also add in math characters such as ≡ø± etc...

I will also be adding in a tray-menu to reveal a help file for the hotkeys included ( and I may set it up so it can be auto generated using simple calls to a library so when you add your own hotkeys you can add help entries for them quickly and easily ), in addition to settings that are enabled / disabled such as language mode ( for if I decide to add in the Spanish keys ¿ etc.. so any that share the same hotkey as the German language such as áéóú can be switched in an instant - I use WIN + F8 ), etc..,

Added Changelog popup / dialog which opens once per new version - config.ini comparison value is under AcecoolAHK_Framework > version_ran which is the last version executed... Still working on this, but the functionality exists so when I get around to designing the changelog dialog window.. it will open :-) - I will also add an updater to let you know when files have been added or changed online and the option to automatically update the files... Note: Because of this feature, I highly recommend you do not edit core files - if you want to add a hotkey then add a new file, etc... If you want to add new functionality to the existing libraries, I'd recommend extending the class for now and renaming the original.. for example... Class Math ... becomes Class __Math .... and you write in Class Math extends __Math .. this way your addition will work by using Math.Function( xxx ) and if you submit a pull-request with your function added then when I update the file it will be renamed back to Math - you will still need to remove your new file ( or you can simply make the changes in the file knowing they will be removed if updater updates- I may add extra functionality later to make the auto-updater more intelligent by taking only the changes and adding them near where they should be in the file ( such as before function x ).... But these won't be in right away.